[refactor] Apply package name change on changed files
