+++ app/src/main/java/ir/composenews/ui/MainViewModel.kt
@@ -1,12 +1,12 @@
-package ir.kaaveh.composenews.ui
+package ir.composenews.ui
 
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import ir.kaaveh.base.BaseViewModel
-import ir.kaaveh.base.MainContract
-import ir.kaaveh.core_test.dispatcher.DispatcherProvider
-import ir.kaaveh.designsystem.utils.ContentType
-import ir.kaaveh.domain.model.Market
+import ir.composenews.base.BaseViewModel
+import ir.composenews.base.MainContract
+import ir.composenews.core_test.dispatcher.DispatcherProvider
+import ir.composenews.domain.model.Market
+import ir.composenews.utils.ContentType
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -16,7 +16,7 @@ import javax.inject.Inject
 @HiltViewModel
 class MainViewModel @Inject constructor(
     dispatcherProvider: DispatcherProvider,
-): BaseViewModel(dispatcherProvider), MainContract {
+) : BaseViewModel(dispatcherProvider), MainContract {
 
     private val mutableState = MutableStateFlow(MainContract.State())
     override val state: StateFlow<MainContract.State> = mutableState.asStateFlow()
@@ -30,7 +30,8 @@ class MainViewModel @Inject constructor(
         mutableState.emit(
             mutableState.value.copy(
                 market,
-                isDetailOnlyOpen = contentType == ContentType.SINGLE_PANE)
+                isDetailOnlyOpen = contentType == ContentType.SINGLE_PANE
+            )
         )
     }
