+++ feature/marketdetail/src/main/java/ir/composenews/marketdetail/MarketDetailScreen.kt
@@ -24,22 +24,13 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import coil.compose.rememberAsyncImagePainter
 import ir.composenews.base.BaseRoute
-import ir.composenews.base.BaseViewModel
+import ir.composenews.base.use
 import ir.composenews.designsystem.component.FavoriteIcon
 import ir.composenews.designsystem.component.QuadLineChart
 import ir.composenews.designsystem.preview.ThemePreviews
 import ir.composenews.designsystem.theme.ComposeNewsTheme
-import ir.composenews.base.use
 import ir.composenews.domain.model.Market
 import ir.composenews.marketdetail.preview_provider.MarketDetailStateProvider
-import ir.kaaveh.base.BaseRoute
-import ir.kaaveh.base.use
-import ir.kaaveh.designsystem.component.FavoriteIcon
-import ir.kaaveh.designsystem.component.QuadLineChart
-import ir.kaaveh.designsystem.preview.ThemePreviews
-import ir.kaaveh.designsystem.theme.ComposeNewsTheme
-import ir.kaaveh.domain.model.Market
-import ir.kaaveh.marketdetail.preview_provider.MarketDetailStateProvider
 
 @Composable
 fun MarketDetailRoute(
