+++ feature/marketlist/src/main/java/ir/composenews/marketlist/MarketListScreen.kt
@@ -22,13 +22,14 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.hilt.navigation.compose.hiltViewModel
 import ir.composenews.base.BaseRoute
-import ir.composenews.base.BaseViewModel
+import ir.composenews.base.MainContract
 import ir.composenews.base.use
 import ir.composenews.designsystem.preview.ThemePreviews
 import ir.composenews.designsystem.theme.ComposeNewsTheme
 import ir.composenews.domain.model.Market
 import ir.composenews.marketlist.component.MarketListItem
 import ir.composenews.marketlist.preview_provider.MarketListStateProvider
+import ir.composenews.utils.ContentType
 
 @Composable
 fun MarketListRoute(
