+++ gradle/libs.versions.toml
@@ -52,6 +52,9 @@ kotlin = "1.9.10"
 
 kotlinxImmutableCollection = "0.3.5"
 
+windowsSizeClass = "1.1.1"
+adaptive = "0.26.2-beta"
+
 [libraries]
 androidx-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "ktx" }
 
@@ -104,6 +107,9 @@ startup-runtime = { group = "androidx.startup", name = "startup-runtime", versio
 work-runtime-ktx = { group = "androidx.work", name = "work-runtime-ktx", version.ref = "workRuntimeKtx" }
 
 kotlinx-collections-immutable = { group = "org.jetbrains.kotlinx", name = "kotlinx-collections-immutable", version.ref = "kotlinxImmutableCollection" }
+
+window-size = { group = "androidx.compose.material3", name = "material3-window-size-class", version.ref = "windowsSizeClass"}
+accomponist-adaptive = { group = "com.google.accompanist", name = "accompanist-adaptive", version.ref = "adaptive"}
 
 # Dependencies of the included build-logic
 kotlin-gradlePlugin = { group = "org.jetbrains.kotlin", name = "kotlin-gradle-plugin", version.ref = "kotlin" }
