+++ library/navigation/src/main/java/ir/composenews/navigation/graph/NewsDetail.kt
@@ -6,14 +6,14 @@ import androidx.compose.animation.core.EaseOut
 import androidx.compose.animation.core.tween
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import ir.composenews.base.BaseViewModel
+import ir.composenews.base.MainContract
 import ir.composenews.domain.model.Market
 import ir.composenews.marketdetail.MarketDetailRoute
 import ir.composenews.navigation.Destinations
 import ir.composenews.navigation.extension_function.parcelableData
 
 fun NavGraphBuilder.marketDetail(
-    onProvideBaseViewModel: (baseViewModel: BaseViewModel) -> Unit,
+    uiState: MainContract.State,
 ) {
     composable(
         route = Destinations.MarketDetailScreen().route,
@@ -31,11 +31,9 @@ fun NavGraphBuilder.marketDetail(
         }
     ) { entry ->
         val market = entry.parcelableData<Market>(Destinations.MarketDetailScreen().market)
+            ?: uiState.market as Market
         MarketDetailRoute(
             market = market,
-            onProvideBaseViewModel = {
-                onProvideBaseViewModel(it)
-            },
         )
     }
 }
